Incorporate NFT staking hooks into dao-hooks package
