+++ contracts/voting/dao-voting-cw721-staked/src/lib.rs
@@ -2,7 +2,6 @@
 
 pub mod contract;
 mod error;
-pub mod hooks;
 pub mod msg;
 pub mod state;
