+++ contracts/voting/dao-voting-cw721-staked/src/testing/mod.rs
@@ -1,5 +1,6 @@
 mod adversarial;
 mod execute;
+mod hooks;
 mod instantiate;
 mod queries;
 mod tests;
