+++ packages/dao-hooks/src/lib.rs
@@ -1,7 +1,7 @@
 #![doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/README.md"))]
 
+mod all_hooks;
+pub mod nft_stake;
 pub mod proposal;
 pub mod stake;
 pub mod vote;
-
-// TODO DAO hooks enum with test that contains all hooks
